+++ internal/client/rebind/transport_test_compat_test.go
@@ -1,9 +1,0 @@
-package rebind
-
-import "github.com/park285/iris-client-go/v2/internal/client/transport"
-
-const PathReply = transport.PathReply
-
-var WithHTTPClient = transport.WithHTTPClient
-var WithTransport = transport.WithTransport
-var WithH3CACertFile = transport.WithH3CACertFile